grant-matching-app/requirements.txt
flask
azure-identity
azure-storage-blob
openai
azure-search-documents
